Fix token balance bug
